[serve] metrics pusher improvements and refactor (#42510)
* Make it possible to re-register / update tasks in metrics pusher.
* move metric spusher and `InMemoryMetricsStore` to `_private/metrics_utils.py`
* move around tests
  * remove `test_autoscaling_metrics.py`. The tests have been split in half: in memory metrics store tests moved to `unit/test_metrics_utils.py`, and the single e2e mmetrics test has been moved to `test_autoscaling_policy.py`.
  * move metrics pusher tests from `test_util.py` to `unit/test_metrics_utils.py`.
