--- python/ray/serve/_private/autoscaling_metrics.py
@@ -1,97 +1,0 @@
-import bisect
-import logging
-from collections import defaultdict
-from dataclasses import dataclass, field
-from typing import DefaultDict, Dict, List, Optional
-
-from ray.serve._private.constants import SERVE_LOGGER_NAME
-
-logger = logging.getLogger(SERVE_LOGGER_NAME)
-
-
-@dataclass(order=True)
-class TimeStampedValue:
-    timestamp: float
-    value: float = field(compare=False)
-
-
-class InMemoryMetricsStore:
-    """A very simple, in memory time series database"""
-
-    def __init__(self):
-        self.data: DefaultDict[str, List[TimeStampedValue]] = defaultdict(list)
-
-    def add_metrics_point(self, data_points: Dict[str, float], timestamp: float):
-        """Push new data points to the store.
-
-        Args:
-            data_points: dictionary containing the metrics values. The
-              key should be a string that uniquely identifies this time series
-              and to be used to perform aggregation.
-            timestamp: the unix epoch timestamp the metrics are
-              collected at.
-        """
-        for name, value in data_points.items():
-            # Using in-sort to insert while maintaining sorted ordering.
-            bisect.insort(a=self.data[name], x=TimeStampedValue(timestamp, value))
-
-    def _get_datapoints(self, key: str, window_start_timestamp_s: float) -> List[float]:
-        """Get all data points given key after window_start_timestamp_s"""
-
-        datapoints = self.data[key]
-
-        idx = bisect.bisect(
-            a=datapoints,
-            x=TimeStampedValue(
-                timestamp=window_start_timestamp_s, value=0  # dummy value
-            ),
-        )
-        return datapoints[idx:]
-
-    def window_average(
-        self, key: str, window_start_timestamp_s: float, do_compact: bool = True
-    ) -> Optional[float]:
-        """Perform a window average operation for metric `key`
-
-        Args:
-            key: the metric name.
-            window_start_timestamp_s: the unix epoch timestamp for the
-              start of the window. The computed average will use all datapoints
-              from this timestamp until now.
-            do_compact: whether or not to delete the datapoints that's
-              before `window_start_timestamp_s` to save memory. Default is
-              true.
-        Returns:
-            The average of all the datapoints for the key on and after time
-            window_start_timestamp_s, or None if there are no such points.
-        """
-        points_after_idx = self._get_datapoints(key, window_start_timestamp_s)
-
-        if do_compact:
-            self.data[key] = points_after_idx
-
-        if len(points_after_idx) == 0:
-            return
-        return sum(point.value for point in points_after_idx) / len(points_after_idx)
-
-    def max(self, key: str, window_start_timestamp_s: float, do_compact: bool = True):
-        """Perform a max operation for metric `key`.
-
-        Args:
-            key: the metric name.
-            window_start_timestamp_s: the unix epoch timestamp for the
-              start of the window. The computed average will use all datapoints
-              from this timestamp until now.
-            do_compact: whether or not to delete the datapoints that's
-              before `window_start_timestamp_s` to save memory. Default is
-              true.
-        Returns:
-            Max value of the data points for the key on and after time
-            window_start_timestamp_s, or None if there are no such points.
-        """
-        points_after_idx = self._get_datapoints(key, window_start_timestamp_s)
-
-        if do_compact:
-            self.data[key] = points_after_idx
-
-        return max((point.value for point in points_after_idx), default=None)
