+++ python/ray/serve/_private/utils.py
@@ -3,11 +3,9 @@ import copy
 import importlib
 import inspect
 import logging
-import math
 import os
 import random
 import string
-import threading
 import time
 import traceback
 from abc import ABC, abstractmethod
@@ -393,106 +391,6 @@ def extract_self_if_method_call(args: List[Any], func: Callable) -> Optional[obj
                 return args[0]
 
     return None
-
-
-class _MetricTask:
-    def __init__(self, task_func, interval_s, callback_func):
-        """
-        Args:
-            task_func: a callable that MetricsPusher will try to call in each loop.
-            interval_s: the interval of each task_func is supposed to be called.
-            callback_func: callback function is called when task_func is done, and
-                the result of task_func is passed to callback_func as the first
-                argument, and the timestamp of the call is passed as the second
-                argument.
-        """
-        self.task_func: Callable = task_func
-        self.interval_s: float = interval_s
-        self.callback_func: Callable[[Any, float]] = callback_func
-        self.last_ref: Optional[ray.ObjectRef] = None
-        self.last_call_succeeded_time: Optional[float] = time.time()
-
-
-class MetricsPusher:
-    """
-    Metrics pusher is a background thread that run the registered tasks in a loop.
-    """
-
-    def __init__(
-        self,
-    ):
-        self.tasks: List[_MetricTask] = []
-        self.pusher_thread: Union[threading.Thread, None] = None
-        self.stop_event = threading.Event()
-
-    def register_task(self, task_func, interval_s, process_func=None):
-        self.tasks.append(_MetricTask(task_func, interval_s, process_func))
-
-    def start(self):
-        """Start a background thread to run the registered tasks in a loop.
-
-        We use this background so it will be not blocked by user's code and ensure
-        consistently metrics delivery. Python GIL will ensure that this thread gets
-        fair timeshare to execute and run.
-        """
-
-        def send_forever():
-            while True:
-                if self.stop_event.is_set():
-                    return
-
-                start = time.time()
-                for task in self.tasks:
-                    try:
-                        if start - task.last_call_succeeded_time >= task.interval_s:
-                            if task.last_ref:
-                                ready_refs, _ = ray.wait([task.last_ref], timeout=0)
-                                if len(ready_refs) == 0:
-                                    continue
-                            data = task.task_func()
-                            task.last_call_succeeded_time = time.time()
-                            if task.callback_func and ray.is_initialized():
-                                task.last_ref = task.callback_func(
-                                    data, send_timestamp=time.time()
-                                )
-                    except Exception as e:
-                        logger.warning(
-                            f"MetricsPusher thread failed to run metric task: {e}"
-                        )
-
-                # For all tasks, check when the task should be executed
-                # next. Sleep until the next closest time.
-                least_interval_s = math.inf
-                for task in self.tasks:
-                    time_until_next_push = task.interval_s - (
-                        time.time() - task.last_call_succeeded_time
-                    )
-                    least_interval_s = min(least_interval_s, time_until_next_push)
-
-                time.sleep(max(least_interval_s, 0))
-
-        if len(self.tasks) == 0:
-            raise ValueError("MetricsPusher has zero tasks registered.")
-
-        self.pusher_thread = threading.Thread(target=send_forever)
-        # Making this a daemon thread so it doesn't leak upon shutdown, and it
-        # doesn't need to block the replica's shutdown.
-        self.pusher_thread.setDaemon(True)
-        self.pusher_thread.start()
-
-    def __del__(self):
-        self.shutdown()
-
-    def shutdown(self):
-        """Shutdown metrics pusher gracefully.
-
-        This method will ensure idempotency of shutdown call.
-        """
-        if not self.stop_event.is_set():
-            self.stop_event.set()
-
-        if self.pusher_thread:
-            self.pusher_thread.join()
 
 
 def call_function_from_import_path(import_path: str) -> Any:
